pokemon_club/requirements.txt
flask
flask_cors
nebula2-python
google-cloud-secret-manager
gunicorn
